refactor(p2p): numerous p2p fixes
Work on testground has unveiled a number of problems that need addressing in our
p2p library. The biggest problem is *still* with UpgradeToQriConnection, which
is firing at the wrong time in a connection lifecycle.

We want to check all new connections for qri support, and upgrade them if we
detect a p2p peer supports Qri. There is a better way to do this, but it's in a
future version of libp2p we don't depend on yet. We're in the process of
upgrading both libp2p *and* IPFS, and should wait for those to land before
moving over to additional refactoring. For now the p2p tests are going to be
*extremely* flaky.

But this commit brings a bunch of cleanup that I think is worth it. namely:

* switch to /p2p/ peer addresses, matching shift coming from IPFS
* remove erroneous context field from QriNode, node.GoOnline now reqires a
  context instead, which is where we should be using contexts
* remove msgChan field from QriNode. It was only being used to write to the list
  of receivers, which is now done directly when we get a message
* only create a discovery service if we're using our own libp2p host (not IPFS)
* adjust discovery handler to only attempt to connect on discovery, matching
  IPFS behaviour. Qri protocol upgrade is handled by network notifee handler
* move "UpgradeToQriConnection" into a network notifee handler. Now all new
  connections will get a uniform check for Qri protocol upgrade
* drop commented-out test: TestRequestProfileOneWayConnection
* add events for peer connect / disconnect